python/relatorios/requirements.txt
psycopg2-binary>=2.9.0
pandas>=2.0.0
reportlab>=4.0.0
matplotlib>=3.7.0
plotly>=5.0.0
python-dotenv>=0.19.0
numpy>=1.24.0
Pillow>=10.0.0 